+++ src/pydvl/utils/types.py
@@ -2,6 +2,7 @@
 transformations. Some of it probably belongs elsewhere.
 """
 import inspect
+from enum import Enum
 from typing import Callable, Optional, Protocol, Type, Union
 
 from numpy import ndarray