+++ src/pydvl/value/__init__.py
@@ -4,3 +4,7 @@ Algorithms for the exact and approximate computation of value and semi-value.
 See :ref:`data valuation` for an introduction to the concepts and methods
 implemented here.
 """
+
+from .results import *  # isort: skip
+from .loo import *
+from .shapley import *