+++ src/pydvl/value/loo/naive.py
@@ -1,26 +1,31 @@
-from typing import OrderedDict
-
 import numpy as np
 
-from pydvl.reporting.scores import sort_values
 from pydvl.utils import Utility, maybe_progress
+from pydvl.value.results import ValuationResult, ValuationStatus
 
 __all__ = ["naive_loo"]
 
 
-def naive_loo(u: Utility, *, progress: bool = True) -> OrderedDict[int, float]:
+def naive_loo(u: Utility, *, progress: bool = True) -> ValuationResult:
     """Computes leave one out score.
 
     No caching nor parallelization is implemented.
 
     :param u: Utility object with model, data, and scoring function
     :param progress: If True, display a progress bar
+    :return: Object with the data values.
     """
 
-    values = {i: 0.0 for i in u.data.indices}
+    values = np.zeros_like(u.data.indices)
     all_indices = set(u.data.indices)
     for i in maybe_progress(data.indices, progress):  # type: ignore
         subset = all_indices.difference({i})
         values[i] = u(u.data.indices) - u(subset)
 
-    return sort_values(values)
+    return ValuationResult(
+        algorithm="naive_loo",
+        status=ValuationStatus.Converged,
+        values=values,
+        stderr=None,
+        data_names=u.data.data_names,
+    )