+++ src/pydvl/value/shapley/actor.py
@@ -16,6 +16,7 @@ from pydvl.utils import Utility, get_running_avg_variance, maybe_progress
 from pydvl.utils.config import ParallelConfig
 from pydvl.utils.parallel.actor import Coordinator, RayActorWrapper, Worker
 from pydvl.utils.parallel.backend import init_parallel_backend
+from pydvl.value.results import ValuationStatus
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
@@ -79,6 +80,7 @@ class ShapleyCoordinator(Coordinator):
             )
         self.value_tolerance = value_tolerance
         self.max_iterations = max_iterations
+        self._status = ValuationStatus.Pending
 
     def get_results(self) -> Tuple["NDArray", "NDArray"]:
         """Aggregates the results of the different workers
@@ -117,11 +119,14 @@ class ShapleyCoordinator(Coordinator):
         """Checks whether the accuracy of the calculation or the total number
         of iterations have crossed the set thresholds.
 
-        If the threshold has been reached, sets the flag
-        :attr:`~ShapleyCoordinator.is_done` to `True`.
+        If any of the thresholds have been reached, then calls to
+        :meth:`~Coordinator.is_done` return `True`.
 
-        :return: value of :attr:`~ShapleyCoordinator.is_done`
+        :return: True if converged or reached max iterations.
         """
+        if self._is_done:
+            return True
+
         if len(self.workers_results) == 0:
             logger.info("No worker has updated its status yet.")
             self._is_done = False
@@ -133,12 +138,19 @@ class ShapleyCoordinator(Coordinator):
                 and std_to_val_ratio < self.value_tolerance
             ):
                 self._is_done = True
+                logger.info("Converged")
+                self._status = ValuationStatus.Converged
             if (
                 self.max_iterations is not None
                 and self._total_iterations > self.max_iterations
             ):
                 self._is_done = True
+                logger.info(f"Max iterations ({self.max_iterations}) reached")
+                self._status = ValuationStatus.MaxIterations
         return self._is_done
+
+    def status(self) -> ValuationStatus:
+        return self._status
 
 
 class ShapleyWorker(Worker):