+++ src/pydvl/value/shapley/naive.py
@@ -1,11 +1,9 @@
 import math
 import warnings
-from collections import OrderedDict
 from itertools import permutations
 
 import numpy as np
 
-from pydvl.reporting.scores import sort_values
 from pydvl.utils import (
     MapReduceJob,
     ParallelConfig,
@@ -14,13 +12,12 @@ from pydvl.utils import (
     maybe_progress,
     powerset,
 )
+from pydvl.value.results import ValuationResult, ValuationStatus
 
 __all__ = ["permutation_exact_shapley", "combinatorial_exact_shapley"]
 
 
-def permutation_exact_shapley(
-    u: Utility, *, progress: bool = True
-) -> "OrderedDict[str, float]":
+def permutation_exact_shapley(u: Utility, *, progress: bool = True) -> ValuationResult:
     r"""Computes the exact Shapley value using the formulation with permutations:
 
     $$v_u(x_i) = \frac{1}{n!} \sum_{\sigma \in \Pi(n)} [u(\sigma_{i-1} \cup {i}) − u(\sigma_{i})].$$
@@ -57,7 +54,13 @@ def permutation_exact_shapley(
             values[idx] += u(p[: i + 1]) - u(p[:i])
     values /= math.factorial(n)
 
-    return sort_values({u.data.data_names[i]: v for i, v in enumerate(values)})
+    return ValuationResult(
+        algorithm="permutation_exact_shapley",
+        status=ValuationStatus.Converged,
+        values=values,
+        stderr=None,
+        data_names=u.data.data_names,
+    )
 
 
 def _combinatorial_exact_shapley(
@@ -89,7 +92,7 @@ def combinatorial_exact_shapley(
     n_jobs: int = 1,
     config: ParallelConfig = ParallelConfig(),
     progress: bool = False,
-) -> "OrderedDict[str, float]":
+) -> ValuationResult:
     r"""Computes the exact Shapley value using the combinatorial definition.
 
     $$v_u(i) = \frac{1}{n} \sum_{S \subseteq N \setminus \{i\}} \binom{n-1}{ | S | }^{-1} [u(S \cup \{i\}) − u(S)].$$
@@ -107,9 +110,7 @@ def combinatorial_exact_shapley(
     :param config: Object configuring parallel computation, with cluster address,
         number of cpus, etc.
     :param progress: Whether to display progress bars for each job.
-
-    :return: Dictionary of {"index or label": exact_value}, sorted by decreasing
-        value.
+    :return: Object with the data values.
     """
     # Arbitrary choice, will depend on time required, caching, etc.
     if len(u.data) // n_jobs > 20:
@@ -131,4 +132,10 @@ def combinatorial_exact_shapley(
         n_jobs=n_jobs,
     )
     values = map_reduce_job(u.data.indices)[0]
-    return sort_values({u.data.data_names[i]: v for i, v in enumerate(values)})
+    return ValuationResult(
+        algorithm="combinatorial_exact_shapley",
+        status=ValuationStatus.Converged,
+        values=values,
+        stderr=None,
+        data_names=u.data.data_names,
+    )