+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import functools
-from collections import OrderedDict, defaultdict
-from typing import TYPE_CHECKING, Dict, Optional, Sequence, Tuple, Type, Union
+from collections import defaultdict
+from typing import TYPE_CHECKING, Optional, Sequence, Tuple, Type
 
 import numpy as np
 import pytest
@@ -11,9 +11,10 @@ from sklearn.linear_model import LinearRegression
 from sklearn.pipeline import make_pipeline
 from sklearn.preprocessing import MinMaxScaler, PolynomialFeatures
 
-from pydvl.utils import Dataset, MemcachedClientConfig, MemcachedConfig, Scorer, Utility
+from pydvl.utils import Dataset, MemcachedClientConfig, Utility
 from pydvl.utils.numeric import random_matrix_with_condition_number
 from pydvl.utils.parallel import available_cpus
+from pydvl.value import ValuationResult, ValuationStatus
 
 if TYPE_CHECKING:
     from _pytest.config import Config
@@ -334,11 +335,18 @@ def dummy_utility(num_samples: int = 10):
 @pytest.fixture(scope="function")
 def analytic_shapley(num_samples):
     """Scores are i/n, so v(i) = 1/n! Σ_π [U(S^π + {i}) - U(S^π)] = i/n"""
+
     u = dummy_utility(num_samples)
-    exact_values = OrderedDict(
-        {i: i / float(max(u.data.x_train)) for i in u.data.indices}
+    m = float(max(u.data.x_train))
+    values = np.array([i / m for i in u.data.indices])
+    result = ValuationResult(
+        algorithm="exact",
+        values=values,
+        stderr=np.zeros_like(values),
+        data_names=u.data.indices,
+        status=ValuationStatus.Converged,
     )
-    return u, exact_values
+    return u, result
 
 
 class TolerateErrors:
@@ -367,34 +375,35 @@ class TolerateErrors:
 
 
 def check_total_value(
-    u: Utility, values: OrderedDict, rtol: float = 0.05, atol: float = 1e-6
+    u: Utility, values: ValuationResult, rtol: float = 0.05, atol: float = 1e-6
 ):
     """Checks absolute distance between total and added values.
     Shapley value is supposed to fulfill the total value axiom."""
     total_utility = u(u.data.indices)
-    values = np.fromiter(values.values(), dtype=float, count=len(u.data))
-    # We could want relative tolerances here if we didn't have the range of
-    # the scorer.
-    assert np.isclose(values.sum(), total_utility, rtol=rtol, atol=atol)
+    # We can use relative tolerances if we don't have the range of the scorer.
+    assert np.isclose(np.sum(values.values), total_utility, rtol=rtol, atol=atol)
 
 
-def check_exact(values: OrderedDict, exact_values: OrderedDict, atol: float = 1e-6):
+def check_exact(
+    values: ValuationResult,
+    exact_values: ValuationResult,
+    rtol: float = 0.1,
+    atol: float = 1e-6,
+):
     """Compares ranks and values."""
 
-    k = list(values.keys())
-    ek = list(exact_values.keys())
+    values.sort()
+    exact_values.sort()
 
-    assert np.all(k == ek), "Ranks do not match"
-
-    v = np.array(list(values.values()))
-    ev = np.array(list(exact_values.values()))
-
-    assert np.allclose(v, ev, atol=atol), f"{v} != {ev}"
+    assert np.all(values.indices == exact_values.indices), "Ranks do not match"
+    assert np.allclose(
+        values.values, exact_values.values, rtol=rtol, atol=atol
+    ), "Values do not match"
 
 
 def check_values(
-    values: Dict,
-    exact_values: Dict,
+    values: ValuationResult,
+    exact_values: ValuationResult,
     rtol: float = 0.1,
     atol: float = 1e-5,
 ):
@@ -415,15 +424,12 @@ def check_values(
         elements in `exact_values`. E.g. if atol = 0.1, and rtol = 0 we must
         have |value - exact_value| < 0.1 for every value.
     """
-    for key in values:
-        assert (
-            abs(values[key] - exact_values[key]) < abs(exact_values[key]) * rtol + atol
-        )
+    assert np.allclose(values.values, exact_values.values, rtol=rtol, atol=atol)
 
 
 def check_rank_correlation(
-    values: OrderedDict,
-    exact_values: OrderedDict,
+    values: ValuationResult,
+    exact_values: ValuationResult,
     k: int = None,
     threshold: float = 0.9,
 ):
@@ -439,18 +445,17 @@ def check_rank_correlation(
         succeed
     """
     # FIXME: estimate proper threshold for spearman
-    if k is not None:
-        raise NotImplementedError
-    else:
-        k = len(values)
-    ranks = np.array(list(values.keys())[:k])
-    ranks_exact = np.array(list(exact_values.keys())[:k])
 
-    correlation, pvalue = spearmanr(ranks, ranks_exact)
+    k = k or len(values)
+
+    values.sort()  # Probably a noop since default is to sort
+    exact_values.sort()
+
+    top_k = np.array([it.index for it in values[-k:]])
+    top_k_exact = np.array([it.index for it in exact_values[-k:]])
+
+    correlation, pvalue = spearmanr(top_k, top_k_exact)
     assert correlation >= threshold, f"{correlation} < {threshold}"
-
-
-# start_logging_server()
 
 
 # Tolerate Errors Plugin
@@ -577,10 +582,7 @@ class TolerateErrorFixture:
             self.session.set_exceptions_to_ignore(self.name, exceptions_to_ignore)
 
     def __call__(
-        self,
-        max_failures: int,
-        *,
-        exceptions_to_ignore: EXCEPTIONS_TYPE = None,
+        self, max_failures: int, *, exceptions_to_ignore: EXCEPTIONS_TYPE = None
     ):
         self.session.set_max_failures(self.name, max_failures)
         self.session.set_exceptions_to_ignore(self.name, exceptions_to_ignore)
@@ -621,9 +623,7 @@ def wrap_pytest_function(pyfuncitem: pytest.Function):
 
 @pytest.fixture(scope="function")
 def tolerate(request: pytest.FixtureRequest):
-    fixture = TolerateErrorFixture(
-        request.node,
-    )
+    fixture = TolerateErrorFixture(request.node)
     return fixture