+++ tests/shapley/test_naive.py
@@ -5,7 +5,10 @@ import pytest
 from sklearn.linear_model import LinearRegression
 
 from pydvl.utils import GroupedDataset, MemcachedConfig, Utility
-from pydvl.value.shapley import combinatorial_exact_shapley, permutation_exact_shapley
+from pydvl.value.shapley.naive import (
+    combinatorial_exact_shapley,
+    permutation_exact_shapley,
+)
 from tests.conftest import check_total_value, check_values
 
 log = logging.getLogger(__name__)
@@ -59,11 +62,7 @@ def test_linear(
 
 @pytest.mark.parametrize(
     "a, b, num_points, num_groups, scorer",
-    [
-        (2, 0, 50, 3, "r2"),
-        (2, 1, 100, 5, "r2"),
-        (2, 1, 100, 5, "explained_variance"),
-    ],
+    [(2, 0, 50, 3, "r2"), (2, 1, 100, 5, "r2"), (2, 1, 100, 5, "explained_variance")],
 )
 def test_grouped_linear(
     linear_dataset,
@@ -117,10 +116,9 @@ def test_linear_with_outlier(
         cache_options=MemcachedConfig(client_config=memcache_client_config),
     )
     shapley_values = permutation_exact_shapley(linear_utility, progress=False)
-    log.info(f"Shapley values: {shapley_values}")
     check_total_value(linear_utility, shapley_values, atol=total_atol)
 
-    assert int(list(shapley_values.keys())[0]) == outlier_idx
+    assert shapley_values.indices[0] == outlier_idx
 
 
 @pytest.mark.parametrize(
@@ -182,7 +180,6 @@ def test_polynomial_with_outlier(
     )
 
     shapley_values = permutation_exact_shapley(poly_utility, progress=False)
-    log.info(f"Shapley values: {shapley_values}")
     check_total_value(poly_utility, shapley_values, atol=total_atol)
 
-    assert int(list(shapley_values.keys())[0]) == outlier_idx
+    assert shapley_values[0].index == outlier_idx